requirements.txt
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
gkeepapi==0.13.4
gpsoauth==1.0.0
idna==2.10
oauthlib==3.1.0
pycryptodomex==3.10.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
schoolopy==0.2.1
six==1.16.0
urllib3==1.26.4
